Begging of 2/15/25 Meeting
