--- src/main/java/frc/robot/subsystems/VisionSubsystemNEW.java
@@ -1,73 +1,0 @@
-package frc.robot.subsystems;
-
-import java.util.List;
-
-import org.photonvision.PhotonCamera;
-import org.photonvision.proto.Photon;
-import org.photonvision.targeting.PhotonPipelineResult;
-import org.photonvision.targeting.PhotonTrackedTarget;
-
-import edu.wpi.first.wpilibj2.command.SubsystemBase;
-import frc.robot.Constants.VisionConstants;
-
-
-public class VisionSubsystemNEW extends SubsystemBase {
-    private PhotonCamera camera;
-    static List<PhotonPipelineResult> cameraResult;
-
-    static PhotonTrackedTarget bestTarget;
-
-    private static boolean isTagDetected;
-
-            
-    public void Vision() {
-        // visionInit(camera);
-        camera = new PhotonCamera(VisionConstants.cameraName);
-
-    }
-
-
-    private void visionInit(PhotonCamera camera) {
-        camera = new PhotonCamera(VisionConstants.cameraName);
-
-    }
-
-    @Override
-    public void periodic() {
-        cameraResult = camera.getAllUnreadResults();
-        if (!cameraResult.isEmpty()) {
-            isTagDetected = true;
-        } else {
-            isTagDetected = false;
-        }
-    }
-
-    /**
-     * Returns the latest camera result
-     * @return List<PhotonPipelineResult>
-     */
-    public static List<PhotonPipelineResult> returnCameraResults() {
-        return cameraResult;
-    }
-
-    /**
-     * <p>
-     * Calculates the best target based on the latest result from the pereodic method in this subsystem.
-     * Returns the best target
-     * <p>
-     * @return PhotonTrackedTarget
-     */
-    public static PhotonTrackedTarget returnBestTarget() {
-        var result = cameraResult.get(cameraResult.size() - 1);
-        bestTarget = result.getBestTarget();
-        return bestTarget;
-    }
-
-    /**
-     * Returns true if a tag is detected, and false if one isn't
-     * @return boolean
-     */
-    public static boolean isTagDetedted() {
-        return isTagDetected;
-    }
-}
